pyproject.toml
[tool.poetry]
name = "tz"
version = "0.1.0"
description = ""
authors = ["m"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
fastapi = "^0.112.0"
pydantic = "^2.8.2"
uvicorn = "^0.30.5"
pydantic-settings = "^2.4.0"
requests-async = "^0.6.2"
beautifulsoup4 = "^4.12.3"
selenium = "^4.23.1"
requests-html = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
